complete

+++ project/src/Controllers/CommentController.php
@@ -5,6 +5,7 @@ namespace src\Controllers;
 use Exceptions\NotFoundException;
 use src\View\View;
 use src\Models\Comments\Comment;
+use src\Models\Users\User;
 // use src\Models\Articles\Article;
 use DateTime;
 
@@ -43,7 +44,8 @@ class CommentController
         // Создаем новый объект комментария
         $comment = new Comment();
         $comment->setArticleId($articleId);
-        $comment->setAuthorId($userId);
+        $user = User::getById(1); // 1 — это admin
+        $comment->setAuthorId($user->getId());
         $comment->setText($commentText);
         $comment->setCreatedAt((new DateTime())->format('Y-m-d H:i:s')); // Устанавливаем текущую дату и время
 
@@ -51,7 +53,7 @@ class CommentController
         $comment->save();
 
         // Перенаправляем пользователя на страницу статьи с якорем для нового комментария
-        header("Location: /articles/$articleId#comment-" . $comment->getId());
+        header("Location: /php/project/www/article/$articleId");
         exit;
     }
 
@@ -63,7 +65,7 @@ class CommentController
             throw new NotFoundException();
         }
 
-        $this->view->renderHtml('/comment/edit.php', ['comment' => $comment]);
+        $this->view->renderHtml('comment/edit', ['comment' => $comment]);
     }
 
     public function update(int $commentId)
@@ -78,18 +80,30 @@ class CommentController
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $comment->setText($_POST['text']);
             $comment->save();
-            header("Location: /articles/$articleId#comment-" . $comment->getId());
+            header("Location: /php/project/www/article/$articleId");
             exit();
         } else{
-            $this->view->renderHtml('/comment/edit.php', ['comment' => $comment]);
+            $this->view->renderHtml('comment/edit', ['comment' => $comment]);
         }
 
     }
 
-        public function delete(int $commentId){
+    public function delete(int $commentId)
+    {
+        // Получаем комментарий по ID
         $comment = Comment::getById($commentId);
+        if ($comment === null) {
+            throw new NotFoundException('Комментарий не найден');
+        }
+    
+        // Получаем ID статьи, к которой относится комментарий
         $articleId = $comment->getArticleId();
+    
+        // Удаляем комментарий
         $comment->delete();
-        header("Location: /articles/$articleId");
+    
+        // Перенаправляем пользователя на страницу статьи
+        header("Location: /php/project/www/article/$articleId");
+        exit();
     }
 }
